various changes
 - Update mojito to 0.96
 - Use multistage build to speed-up builds

+++ full/mojito-webapp-launcher.sh
@@ -3,9 +3,11 @@
 set -e
 cd /opt/mojito
 
-# Env vars interpolation
-envsubst < application.properties.dist > application.properties
-rm application.properties.dist
+# Env vars interpolation (if not already processed)
+if [ -f application.properties.dist ]; then
+    envsubst < application.properties.dist > application.properties
+    rm application.properties.dist
+fi
 
 # Go!
 mkdir -p /opt/mojito/workdir && cd /opt/mojito/workdir
